Build RedemptionCode and Redemption models.
AppSumo requires that merchants generate unique redemption codes that can be used to get access to the application they purchased the code for.

The plan here is to use the `RedemptionCode` model to store a unique value that wil end up being the redemption code sold on AppSumo. The `RedemptionCode` model will also be associated with a `Plan` so that we can grant a `User` `Pay::Subscription`.

The `Redemption` model connects the `User` and the `RedemptionCode` and prevents duplicates through a uniqueness scope.

I'm not totally sure if there's a database constraint preventing a `User` from having multiple `Redemption` records, but I use a `has_one` relationship.