Add ability to register with a redemption code. (#176)
* Build RedemptionCode and Redemption models.

AppSumo requires that merchants generate unique redemption codes that can be used to get access to the application they purchased the code for.

The plan here is to use the `RedemptionCode` model to store a unique value that wil end up being the redemption code sold on AppSumo. The `RedemptionCode` model will also be associated with a `Plan` so that we can grant a `User` `Pay::Subscription`.

The `Redemption` model connects the `User` and the `RedemptionCode` and prevents duplicates through a uniqueness scope.

I'm not totally sure if there's a database constraint preventing a `User` from having multiple `Redemption` records, but I use a `has_one` relationship.

* Fix association names.

I needed to ensure a User can only have one Redemption.

Grant user a subscription when registering with a redemption code.

I need to consider setting an end date to the subscription, otherwise it will be a lifetime deal.

* Grant user free trial when registering with a redemption code.

The redemption codes are purchased through AppSumo but I can't afford to grant lifetime access with a fake subscription. So instead, we'll just grant users a free trial.

* Prevent a redemption code from being redeemed more than once.

I originally thought that by scoping this between a user and a redemption_code was good enough, but I realized that does not stop a different user from using the same redemption code. This could be an issue of someone shares their code.

* Optimize registration flow when using a redemption code.

I decided to skip user confirmation since they've already gone through the trouble of buying a registration code. I also handle the case where a redemption code is re-used.

* Add ability to grant lifetime access or trial through a redemption code.

It might be beneficial to have a flexible system where I can choose to grant lifetime access through a subscription with a :fake_processor, or set a trial_ends_at value on the user model.

* Build task to create redemption code records.

```bash
rails "generate_redemption_codes:perform[100, 1, YYYYMMDD]"
```

- First parameter is the amount of records to create (required)
- Second parameter is the ID of the plan to associate with (required)
- Third parameter is the date that trial should end (optional)

I made sure to have the script return a list of the values since I'll need these for AppSumo.