Increase thumbnail size.
Description
-----------

This closes #181 which I think is caused by large images needing to be resized into smaller thumbnails. I cannot reproduce this locally, so it's either an issue with the version of ImageMagick or the amount of resources on my server.

It's possible I will need to adjust the variant sizes again after deployment since I cannot test this locally.

Honeybadger
-----------

https://app.honeybadger.io/projects/77796/faults/70895427